Refactor persist and ingest packages
This commit moves the contents of the ingest and persist packages into
the reading packages. To me they make more sense living in there, as
they are directly tied to readings. It will also help match up with
the new statistics package which contains the full stack for
statistics (transport, persistence etc).

Signed-off-by: David Bond <davidsbond93@gmail.com>