Add validation for sensor types
This commit modifies the SQL schema to use a custom postgres
enum to store the sensor type. This will allow us to catch errors
for invalid sensor names when they are persisted. It also updates
the `Reading` type to use a named type for the sensor which can be
one of a handful of valid values.

The ingestion API will now return an array containing all invalid
readings that have been posted to it.

Signed-off-by: David Bond <davidsbond93@gmail.com>