Add API endpoint to power daily charts (#16)
This commit adds a new endpoint to the API that returns bucketed
reading data for a given date. This can be used to power the UI
charts that display daily data. The time buckets are fixed to 15
minute intervals, so there should be some decent visualisation we
can create from this.

It supports all sensor types and will validate the sensor type
on request.

Signed-off-by: David Bond <davidsbond93@gmail.com>